Added ball collision, sound resources
Modified the _physics_process function of ball.gd so it bounces off paddles as well as the top and bottom walls. For the walls it uses a built-in method bounce, which takes a variable collision, created by a method move_and_collide. For the paddles, the direction is set as a vector pointing from the position of the paddle to the position of the ball. Then it calculates the velocity by multiplying normalized direction vector by speed. Also added sound resources for the ball hitting the walls and for the ball hitting the paddles.